+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - `Player.getAudioTrack` and `Player.getSubtitleTrack` APIs to get currently selected audio and subtitle tracks
 - `SourceConfig.description` property to allow setting a description for the source
+- `Player.getThumbnail` and `Source.getThumbnail` APIs to get thumbnail images
 
 ## [0.10.0] (2023-09-04)